lesson3: Sub-task 3: Deployment strategies
To Post service add a new field topic (:String). This is the topic for the post. You can specify it when creating a new post and when updating existing post. This field also should be returned at the responses for POST, PUT and GET operations.
Build a new docker image of application with changes and push it to the Docker Hub (specify another version of container).
Add Rolling-update deployment strategy to your deployments at manifest files and apply the  manifest, so the old versions of microservices are deployed and running.
Set app version of app containers to the new one and apply manifest one more time. Make sure that new changes are deployed.